requirements.txt
langchain
python-dotenv
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette